GUI/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
awscli==1.33.9
Babel==2.13.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.2
black==24.4.2
bleach==6.1.0
boto3==1.34.127
botocore==1.34.127
bottle==0.12.25
branca==0.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
clr-loader==0.2.6
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dnspython==2.6.1
docker==7.1.0
docutils==0.16
effdet @ file:///F:/efficientdet-pytorch
efficientnet==1.1.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
ExifRead==3.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.18.1
ffmpy==0.3.2
filelock==3.14.0
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.44.0
fqdn==1.5.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-pasta==0.2.0
gradio==4.31.5
gradio_client==0.16.4
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imageio==2.33.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
iopath==0.1.9
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==3.3.3
Keras-Applications==1.0.8
keyring==8.7
keyrings.alt==5.0.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
lxml==4.9.3
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.8.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.1
omegaconf==2.3.0
onnx==1.14.0
onnxruntime==1.15.1
onnxruntime-gpu==1.18.0
onnxsim==0.4.36
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.1
orjson==3.10.5
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
Pillow==10.1.0
platformdirs==3.11.0
portalocker==2.8.2
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.3
proxy_tools==0.1.0
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaml==24.4.0
pyasn1==0.6.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.8.0
pydantic_core==2.20.0
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pythonnet==3.0.3
pytz==2023.3.post1
pywebview==5.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.12.0
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.5.0
s3transfer==0.10.1
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scikit-optimize==0.10.2
scipy==1.13.0
seaborn==0.13.0
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.9.26
timm==1.0.7
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1+cu117
torch-tb-profiler==0.4.3
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.12.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.30.1
virtualenv==20.26.2
voluptuous==0.14.2
watchfiles==0.22.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wrapt==1.16.0
xyzservices==2024.6.0
yacs==0.1.8
yolo==0.3.1
